Add API endpoint to return all qpc reports as tar.gz (#1705)
