+++ quipucords/api/deployments_report/tests_deployments_report.py
@@ -15,9 +15,9 @@ import tarfile
 import uuid
 
 from api.common.common_report import create_report_version
+from api.common.report_csv_renderer import ReportCSVRenderer
 from api.common.report_json_gzip_renderer import ReportJsonGzipRenderer
-from api.deployments_report.csv_renderer import (DeploymentCSVRenderer,
-                                                 sanitize_row)
+from api.deployments_report.util import (sanitize_row)
 from api.models import (Credential,
                         ServerInformation,
                         Source)
@@ -267,8 +267,8 @@ class DeploymentReportTest(TestCase):
                          ['data', None, 'data;data'])
 
     def test_csv_renderer(self):
-        """Test DeploymentCSVRenderer."""
-        renderer = DeploymentCSVRenderer()
+        """Test ReportCSVRenderer."""
+        renderer = ReportCSVRenderer()
         # Test no FC id
         test_json = {}
         value = renderer.render(test_json)
@@ -298,8 +298,8 @@ class DeploymentReportTest(TestCase):
             self.assertEqual(result, True)
 
     def test_csv_renderer_only_name(self):
-        """Test DeploymentCSVRenderer name filter."""
-        renderer = DeploymentCSVRenderer()
+        """Test ReportCSVRenderer name filter."""
+        renderer = ReportCSVRenderer()
         # Test no FC id
         test_json = {}
         value = renderer.render(test_json)
